refactor: 공유로 작성한 사람이 있을 경우에만 보여줄 수 있도록 파일 분리 및 조건 설정
